moved the trigger_build out of worker
